Refactor ProjectList and Tag components to simplify filtering logic - Move the filtering logic to the parent component ProjectList - Remove unnecessary filteredProjects state - Move main layout with TaskSuggestion to homepage
